feat: Add RootLayout component
This commit adds a new component called RootLayout to the web application. The RootLayout component is responsible for rendering the header, footer, and the main content of the application. It also includes metadata for the page title and description. This component will be used as the layout for all pages in the application.